+++ sandbox/src/mocks/r4/getService/responses/sampleServiceWithFullAttributes.json
