+++ sandbox/src/mocks/r4/getService/responses/sampleServiceWithMinimumAttributes.json
