+++ sandbox/src/mocks/r4/searchForServices/responses/searchServiceWithMaxAndMinlAttributes.json
