+++ sandbox/src/mocks/r4/searchForServices/responses/searchServiceWithMinmumalAttributes.json
