+++ sandbox/src/routes/index.js
@@ -5,8 +5,8 @@ const createReferralAndSendForTriage = require('./createReferralAndSendForTriage
 const patientServiceSearch = require('./patientServiceSearch')
 const getCodeSystem = require('./getCodeSystem')
 const retrieveAppointmentSlots = require('./retrieveAppointmentSlots')
-const retrieveBusinessFunctions = require('./retrieveBusinessFunctions')
-const retrieveOboUsers = require('./retrieveOboUsers')
+const retrieveBusinessFunctions = require('./r4/retrieveBusinessFunctions')
+const retrieveOboUsers = require('./r4/retrieveOboUsers')
 const uploadFileToDocumentStore = require('./uploadFileToDocumentStore')
 const generatePatientLetter = require('./generatePatientLetter')
 const retrieveAttachment = require('./retrieveAttachment')
@@ -17,8 +17,8 @@ const updateAppointment = require('./updateAppointment')
 const retrieveClinicalInformation = require('./retrieveClinicalInformation')
 const retrieveWorklist = require('./retrieveWorklist')
 const retrieveAdviceAndGuidanceWorklist = require('./retrieveAdviceAndGuidanceWorklist')
-const retrieveHealthcareService = require('./retrieveHealthcareService')
-const searchForHealthcareServices = require('./searchForHealthcareServices')
+const retrieveHealthcareService = require('./r4/retrieveHealthcareService')
+const searchForHealthcareServices = require('./r4/searchForHealthcareServices')
 const retrieveAdviceAndGuidanceRequest = require('./retrieveAdviceAndGuidanceRequest')
 const retrieveAdviceAndGuidanceConversation = require('./retrieveAdviceAndGuidanceConversation')
 const sendAdviceAndGuidanceResponse = require('./sendAdviceAndGuidanceResponse')
