+++ specification/components/r4/schemas/extensions/Extension-eRS-HoursOfOperation.yaml
