+++ specification/components/r4/schemas/headers/request/BearerAuthorization.yaml
@@ -1,0 +1,11 @@
+in: header
+name: Authorization
+description: |
+  An [OAuth 2.0 bearer token](https://digital.nhs.uk/developer/guides-and-documentation/security-and-authorisation#user-restricted-apis).
+
+  Required in all environments except sandbox.
+required: true
+schema:
+  type: string
+  format: '^Bearer\ [[:ascii:]]+$'
+  example: 'Bearer g1112R_ccQ1Ebbb4gtHBP1aaaNM'
