+++ specification/components/r4/schemas/headers/request/OdsCode.yaml
@@ -1,0 +1,8 @@
+in: header
+name: NHSD-End-User-Organisation-ODS
+description: | 
+  The ODS code of the caller's Organisation
+required: true
+schema:
+  type: string
+  example: 'R69'
