+++ specification/components/r4/schemas/headers/response/CorrelationID.yaml
@@ -1,0 +1,6 @@
+description: | 
+  The X-Correlation-ID from the request header, if supplied, mirrored back.
+required: false
+schema:
+  type: string
+  example: '11C46F5F-CDEF-4865-94B2-0EE0EDCC26DA'
