+++ specification/components/r4/schemas/path/ServiceId.yaml
@@ -1,0 +1,8 @@
+name: id
+in: path
+description: |
+    The ID of the service to be retrieved. It must be valid and relate to an existing Healthcare Service otherwise a HTTP 404 response will be returned.
+required: true
+schema:
+  type: string
+  example: '1'
