+++ specification/components/r4/schemas/responses/BadRequest.yaml
@@ -1,0 +1,1 @@
+description: Bad Request
