+++ specification/components/r4/schemas/responses/Forbidden.yaml
@@ -1,0 +1,1 @@
+description: Forbidden
