+++ specification/components/r4/schemas/responses/InternalServerError.yaml
@@ -1,0 +1,1 @@
+description: Internal Server Error
