+++ specification/components/r4/schemas/responses/NotFound.yaml
@@ -1,0 +1,1 @@
+description: Not Found
