+++ specification/components/r4/schemas/responses/Unauthorized.yaml
@@ -1,0 +1,1 @@
+description: Unauthorized
