+++ specification/components/r4/schemas/responses/UnsupportedMediaType.yaml
@@ -1,0 +1,5 @@
+description: Unsupported Media Type
+content:
+  text/plain;charset=utf-8:
+    schema:
+      type: string
