+++ specification/components/r4/schemas/responses/retrieveOboPractitioners/200Response.yaml
