+++ specification/components/schemas/eRS-Communication.yaml
@@ -92,7 +92,7 @@ properties:
             - identifier
           properties:
             identifier:
-              $ref: 'identifiers/R4_Organisation.yaml'
+              $ref: 'identifiers/STU3_ODS_Organisation.yaml'
   sent:
     type: string
     format: date-time
