+++ specification/components/schemas/extensions/Extension-eRS-Commissioning-Rule-Org.yaml
@@ -11,4 +11,4 @@ properties:
      - 'https://fhir.nhs.uk/STU3/StructureDefinition/Extension-eRS-Commissioning-Rule-Org-1'
     example: 'https://fhir.nhs.uk/STU3/StructureDefinition/Extension-eRS-Commissioning-Rule-Org-1'
   valueIdentifier:
-    $ref: '../identifiers/R4_Organisation.yaml'
+    $ref: '../identifiers/STU3_ODS_Organisation.yaml'
