+++ tests/sandbox/r4/test_a030_retrieve_eRS_business_functions.py
@@ -11,29 +11,22 @@ from utils import HttpMethod
 
 
 @pytest.mark.sandbox
-class TestSearchForHealthcareServices(SandboxTest):
-    authorised_actor_data = [Actor.RC, Actor.RC_DEV, Actor.RCA]
-
-    allowed_business_function_data = [
-        "REFERRING_CLINICIAN",
-        "REFERRING_CLINICIAN_ADMIN",
+class TestRetrieveERSBusinessFunctions(SandboxTest):
+    authorised_actor_data = [
+        Actor.SPC,
+        Actor.SPCA,
+        Actor.SPA,
+        Actor.RC,
+        Actor.RC_DEV,
+        Actor.RCA,
+        Actor.RA,
     ]
 
-    testdata = [
-        (
-            "1,2",
-            "searchForServices/responses/searchServiceWithMinmumalAttributes.json",
-        ),
-        (
-            "3,4",
-            "searchForServices/responses/searchServiceWithMaxAndMinlAttributes.json",
-        ),
-        ("5,6", "searchForServices/responses/searchServiceWithEmptyResponse.json",),
-    ]
+    allowed_business_function_data = []
 
     @pytest.fixture
     def endpoint_url(self) -> str:
-        return "FHIR/R4/HealthcareService"
+        return "FHIR/R4/PractitionerRole"
 
     @pytest.fixture
     def http_method(self) -> HttpMethod:
@@ -41,32 +34,31 @@ class TestSearchForHealthcareServices(SandboxTest):
 
     @pytest.fixture
     def authorised_actors(self) -> Iterable[Actor]:
-        return TestSearchForHealthcareServices.authorised_actor_data
+        return TestRetrieveERSBusinessFunctions.authorised_actor_data
 
     @pytest.fixture
     def allowed_business_functions(self) -> Iterable[str]:
-        return TestSearchForHealthcareServices.allowed_business_function_data
+        return TestRetrieveERSBusinessFunctions.allowed_business_function_data
 
     @pytest.fixture
     def call_endpoint(
         self, call_endpoint_url_with_query: Callable[[Actor, Dict[str, str]], Response],
     ) -> Callable[[Actor], Response]:
         return lambda actor, headers={}: call_endpoint_url_with_query(
-            actor, {"_id": "1,2"}, headers,
+            actor, {}, headers,
         )
 
     @pytest.mark.parametrize("actor", authorised_actor_data)
-    @pytest.mark.parametrize("ids,response", testdata)
     def test_success(
         self,
         call_endpoint_url_with_query: Callable[[Actor, Dict[str, str]], Response],
         load_json: Callable[[str], Dict[str, str]],
         actor: Actor,
-        ids,
-        response,
     ):
-        expected_response = load_json(response)
-        actual_response = call_endpoint_url_with_query(actor, {"_id": ids})
+        expected_response = load_json(
+            "r4/retrieveBusinessFunctions/responses/PractitionerRoleBundle.json"
+        )
+        actual_response = call_endpoint_url_with_query(actor)
 
         asserts.assert_status_code(200, actual_response.status_code)
         asserts.assert_response(expected_response, actual_response)
