+++ tests/sandbox/r4/test_a035_search_for_healthcare_services.py
@@ -11,16 +11,29 @@ from utils import HttpMethod
 
 
 @pytest.mark.sandbox
-class TestRetrieveERSBusinessFunctions(SandboxTest):
-    authorised_actor_data = [
-        Actor.SPCA,
+class TestSearchForHealthcareServices(SandboxTest):
+    authorised_actor_data = [Actor.RC, Actor.RC_DEV, Actor.RCA]
+
+    allowed_business_function_data = [
+        "REFERRING_CLINICIAN",
+        "REFERRING_CLINICIAN_ADMIN",
     ]
 
-    allowed_business_function_data = ["SERVICE_PROVIDER_CLINICIAN_ADMIN"]
+    testdata = [
+        (
+            "1,2",
+            "r4/searchForServices/responses/searchServiceWithMinmumalAttributes.json",
+        ),
+        (
+            "3,4",
+            "r4/searchForServices/responses/searchServiceWithMaxAndMinlAttributes.json",
+        ),
+        ("5,6", "r4/searchForServices/responses/searchServiceWithEmptyResponse.json",),
+    ]
 
     @pytest.fixture
     def endpoint_url(self) -> str:
-        return "FHIR/R4/Practitioner"
+        return "FHIR/R4/HealthcareService"
 
     @pytest.fixture
     def http_method(self) -> HttpMethod:
@@ -28,31 +41,32 @@ class TestRetrieveERSBusinessFunctions(SandboxTest):
 
     @pytest.fixture
     def authorised_actors(self) -> Iterable[Actor]:
-        return TestRetrieveERSBusinessFunctions.authorised_actor_data
+        return TestSearchForHealthcareServices.authorised_actor_data
 
     @pytest.fixture
     def allowed_business_functions(self) -> Iterable[str]:
-        return TestRetrieveERSBusinessFunctions.allowed_business_function_data
+        return TestSearchForHealthcareServices.allowed_business_function_data
 
     @pytest.fixture
     def call_endpoint(
         self, call_endpoint_url_with_query: Callable[[Actor, Dict[str, str]], Response],
     ) -> Callable[[Actor], Response]:
         return lambda actor, headers={}: call_endpoint_url_with_query(
-            actor, {"_query": "onBehalfOf"}, headers,
+            actor, {"_id": "1,2"}, headers,
         )
 
     @pytest.mark.parametrize("actor", authorised_actor_data)
+    @pytest.mark.parametrize("ids,response", testdata)
     def test_success(
         self,
         call_endpoint_url_with_query: Callable[[Actor, Dict[str, str]], Response],
         load_json: Callable[[str], Dict[str, str]],
         actor: Actor,
+        ids,
+        response,
     ):
-        expected_response = load_json(
-            "retrieveOboUsers/responses/PractitionerBundle.json"
-        )
-        actual_response = call_endpoint_url_with_query(actor, {"_query": "onBehalfOf"})
+        expected_response = load_json(response)
+        actual_response = call_endpoint_url_with_query(actor, {"_id": ids})
 
         asserts.assert_status_code(200, actual_response.status_code)
         asserts.assert_response(expected_response, actual_response)
