+++ tests/sandbox/r4/test_a037_retrieve_healthcare_service_version.py
@@ -17,8 +17,8 @@ class TestRetrieveHealthcareService(SandboxTest):
     allowed_business_function_data = [""]
 
     testdata = [
-        ("1", "getService/responses/sampleServiceWithMinimumAttributes.json",),
-        ("2", "getService/responses/sampleServiceWithFullAttributes.json",),
+        ("1", "r4/getService/responses/sampleServiceWithMinimumAttributes.json",),
+        ("2", "r4/getService/responses/sampleServiceWithFullAttributes.json",),
     ]
 
     @pytest.fixture
