+++ tests/sandbox/r4/test_a040_retrieve_practioner_info.py
@@ -13,20 +13,14 @@ from utils import HttpMethod
 @pytest.mark.sandbox
 class TestRetrieveERSBusinessFunctions(SandboxTest):
     authorised_actor_data = [
-        Actor.SPC,
         Actor.SPCA,
-        Actor.SPA,
-        Actor.RC,
-        Actor.RC_DEV,
-        Actor.RCA,
-        Actor.RA,
     ]
 
-    allowed_business_function_data = []
+    allowed_business_function_data = ["SERVICE_PROVIDER_CLINICIAN_ADMIN"]
 
     @pytest.fixture
     def endpoint_url(self) -> str:
-        return "FHIR/R4/PractitionerRole"
+        return "FHIR/R4/Practitioner"
 
     @pytest.fixture
     def http_method(self) -> HttpMethod:
@@ -45,7 +39,7 @@ class TestRetrieveERSBusinessFunctions(SandboxTest):
         self, call_endpoint_url_with_query: Callable[[Actor, Dict[str, str]], Response],
     ) -> Callable[[Actor], Response]:
         return lambda actor, headers={}: call_endpoint_url_with_query(
-            actor, {}, headers,
+            actor, {"_query": "onBehalfOf"}, headers,
         )
 
     @pytest.mark.parametrize("actor", authorised_actor_data)
@@ -56,9 +50,9 @@ class TestRetrieveERSBusinessFunctions(SandboxTest):
         actor: Actor,
     ):
         expected_response = load_json(
-            "retrieveBusinessFunctions/responses/PractitionerRoleBundle.json"
+            "r4/retrieveOboUsers/responses/PractitionerBundle.json"
         )
-        actual_response = call_endpoint_url_with_query(actor)
+        actual_response = call_endpoint_url_with_query(actor, {"_query": "onBehalfOf"})
 
         asserts.assert_status_code(200, actual_response.status_code)
         asserts.assert_response(expected_response, actual_response)
